Description
-----
Currently, the default journal items view searches ilike over `account_id` and `partner_id`. This generates subqueries and causes Postgres to use a sequential scan instead of hitting the trigram indexes when OR'd. Instead we can leverage the fact that the number of accounts and partners are usually low compared to the number of aml's, by resolving the many2one domain first and injecting the matched ids into the final domain. We make use of the existing `search_account_id` field and apply the same logic for a new `search_partner_id` field.

Benchmarks
-----
|Counts                              |Before|After|
|------------------------------------|------|-----|
|700k amls, 80k partners, 5k accounts|2.87s |0.45s|
|7M amls, 800k partners, 50k accounts|29s   |4s   |

opw-5047423