@robodoo 2025-09-29
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/229133.png)](https://mergebot.odoo.com/odoo/odoo/pull/229133)