@pivi-odoo 2025-10-01
TODO: Will first be reviewed internally