@juwu-odoo 2025-10-30
Deemed a limitation since the query after the patch still hits the slow account_move_line_date_name_id_idx, it's not really a clear improvement.

The improvement on the index scan time from the patch benchmark seems to be specific to the table data statistics and the searched value. This is corroborated by the fact that some of my benchmarks on values that returned 10,000+ rows show no improvement.